Change CUBLAS precision to double

+++ directsolver/src/cudacg.h
@@ -1,4 +1,4 @@
 #include <vector>
 #include <tuple>
 
-std::tuple<long, long, int> runCusparseCublasCG(std::vector<int> &I, std::vector<int> &J, std::vector<float> &val, std::vector<float> &rhs, std::vector<float> &x, int M, int N, int nz, float tol, int max_iter);
+std::tuple<long, long, int> runCusparseCublasCG(std::vector<int> &I, std::vector<int> &J, std::vector<double> &val, std::vector<double> &rhs, std::vector<double> &x, int M, int N, int nz, double tol, int max_iter);
